Refactor out orientation calculations to seperate class
