+++ android/src/main/java/com/wix/RNCameraKit/camera/CameraViewManager.java
@@ -1,21 +1,17 @@
 package com.wix.RNCameraKit.camera;
 
-import android.app.Activity;
 import android.content.Context;
 import android.graphics.PixelFormat;
 import android.graphics.Point;
 import android.hardware.Camera;
 import android.hardware.SensorManager;
 import android.support.annotation.IntRange;
-import android.util.Log;
 import android.view.Display;
 import android.view.OrientationEventListener;
-import android.view.Surface;
 import android.view.WindowManager;
 
 import com.facebook.react.uimanager.SimpleViewManager;
 import com.facebook.react.uimanager.ThemedReactContext;
-import com.wix.RNCameraKit.DeviceUtils;
 import com.wix.RNCameraKit.Utils;
 
 import java.io.IOException;
@@ -23,10 +19,11 @@ import java.util.List;
 import java.util.Stack;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-@SuppressWarnings("deprecation MagicNumber") // We're still using Camera API 1, everything is deprecated
+import static com.wix.RNCameraKit.camera.Orientation.getSupportedRotation;
+
+@SuppressWarnings("MagicNumber deprecation") // We're still using Camera API 1, everything is deprecated
 public class CameraViewManager extends SimpleViewManager<CameraView> {
 
-    private static final int PORTRAIT_ROTATION = 90;
     private static Camera camera = null;
     private static int currentCamera = 0;
     private static String flashMode = Camera.Parameters.FLASH_MODE_AUTO;
@@ -180,78 +177,14 @@ public class CameraViewManager extends SimpleViewManager<CameraView> {
         Camera.Parameters parameters = camera.getParameters();
         parameters.setRotation(supportedRotation);
         parameters.setPictureFormat(PixelFormat.JPEG);
-        camera.setDisplayOrientation(getDeviceOrientation());
+        camera.setDisplayOrientation(Orientation.getDeviceOrientation(reactContext.getCurrentActivity()));
         camera.setParameters(parameters);
-    }
-
-    private static int getDeviceOrientation() {
-        Activity activity = reactContext.getCurrentActivity();
-        if (activity == null) return PORTRAIT_ROTATION;
-        int rotation = activity.getWindowManager().getDefaultDisplay().getRotation();
-        Camera.CameraInfo info = getCameraInfo();
-        int degrees = 0;
-        switch (rotation) {
-            case Surface.ROTATION_0: degrees = 0; break;
-            case Surface.ROTATION_90: degrees = 90; break;
-            case Surface.ROTATION_180: degrees = 180; break;
-            case Surface.ROTATION_270: degrees = 270; break;
-        }
-
-        int result;
-        if (info.facing == Camera.CameraInfo.CAMERA_FACING_FRONT) {
-            result = (info.orientation + degrees) % 360;
-            result = (360 - result) % 360;  // compensate the mirror
-        } else {  // back-facing
-            result = (info.orientation - degrees + 360) % 360;
-        }
-        return result;
-    }
-
-    private static int getSupportedRotation(int rotation) {
-        int degrees = convertRotationToSupportedAxis(rotation);
-        return isFrontFacingCamera() ? adaptFrontCamera(degrees) : adaptBackCamera(degrees);
-    }
-
-    private static int convertRotationToSupportedAxis(int rotation) {
-        if (rotation < 45) {
-            return 0;
-        } else if (rotation < 135) {
-            return  90;
-        } else if (rotation < 225) {
-            return 180;
-        } else if (rotation < 315){
-            return 270;
-        }
-        return 0;
     }
 
     public static Camera.CameraInfo getCameraInfo() {
         Camera.CameraInfo info = new Camera.CameraInfo();
         Camera.getCameraInfo(currentCamera, info);
         return info;
-    }
-
-    private static boolean isFrontFacingCamera() {
-        return getCameraInfo().facing == Camera.CameraInfo.CAMERA_FACING_FRONT;
-    }
-
-    private static int adaptBackCamera(int degrees) {
-        return (getCameraInfo().orientation - degrees + 360) % 360;
-    }
-
-    private static int adaptFrontCamera(int degrees) {
-        if (DeviceUtils.isGoogleDevice()) {
-            int result = (getCameraInfo().orientation + degrees) % 360;
-            result = (result) % 360;  // compensate the mirror
-            Log.i("CameraViewManager", "adaptFrontCamera for google device: result: [" + result + "] degrees: [" + degrees + "]");
-            return result;
-        } else {
-            // This works on all devices except nexus and pixel
-            int result = (getCameraInfo().orientation + degrees + 180) % 360;
-            result = (result) % 360;  // compensate the mirror
-            Log.i("CameraViewManager", "adaptFrontCamera: result: [" + result + "] degrees: [" + degrees + "]");
-            return result;
-        }
     }
 
     private static Camera.Size getOptimalPreviewSize(List<Camera.Size> sizes, int w, int h) {
